feat: add trigger management functionality and UI components

+++ src/drivers/sqlite/sqlite-generate-schema.ts
@@ -3,6 +3,7 @@ import {
   DatabaseTableColumn,
   DatabaseTableColumnConstraint,
   DatabaseTableSchemaChange,
+  DatabaseTriggerSchemaChange,
 } from "@/drivers/base-driver";
 import { omit, isEqual } from "lodash";
 
@@ -116,11 +117,18 @@ function generateConstraintScript(con: DatabaseTableColumnConstraint) {
 }
 
 export default function generateSqlSchemaChange(
-  change: DatabaseTableSchemaChange
+  change: DatabaseTableSchemaChange & DatabaseTriggerSchemaChange,
+  type: "table" | "trigger" = "table"
 ): string[] {
   const isCreateScript = !change.name.old;
 
   const lines = [];
+
+  if (type === "trigger") {
+    return [
+      `CREATE TRIGGER ${escapeIdentity(change.name.new ?? "")} \n${change.when} ${change.operation} ON ${escapeIdentity(change.tableName)} \nFOR EACH ROW \nBEGIN \n\t${change.statement} \nEND`,
+    ];
+  }
 
   for (const col of change.columns) {
     if (col.new === null) lines.push(`DROP COLUMN ${col.old?.name}`);
